net/ice: fix storing field offset arg
The field offset argument type inside ice driver is of "int" type, but
when parsing and storing the value from a devarg, a uint8_t type is
used, leading to 3 bytes being potentially uninitialized. Fix this by
storing to an int type instead.

Bugzilla ID: 2038
Fixes: df539aaf35cb ("net/ice: refactor flex descriptor protocol field extraction")
Cc: stable@dpdk.org

Acked-by: Anatoly Burakov <anatoly.burakov@intel.com>

+++ drivers/net/intel/ice/ice_ethdev.c
@@ -742,7 +742,7 @@ static int
 handle_field_offs_arg(__rte_unused const char *key, const char *value,
 				void *offs_args)
 {
-	uint8_t *offset = offs_args;
+	int *offset = offs_args;
 
 	if (value == NULL || offs_args == NULL)
 		return -EINVAL;
